+++ CHANGELOG.md
@@ -189,3 +189,4 @@
 - 2023-11-28: Cache ready pool, geo, discovery, daemon, +6
 - 2023-11-29: Builder mod, bettercontact, configuration, discover, +2
 - 2023-11-29: Impl campaigns, mod, leads, qualify lead, +6 query
+- 2023-11-30: Bench chat, task, campaigns, qualify lead, +6